Cleanup regex in saveApiToken
More fail-save approach to parse the URL. The current regex assumes the code and state parameters always appear in the URL in the same order, and that there is no other parameter with a name ending in "code". As the state parameter value is not used at all, not looking for it is slightly more efficient.